- add object tracking to command system

+++ gobot_msg_srv/include/gobot_msg_srv/set_robot_class.h
@@ -30,7 +30,7 @@ namespace robot_class {
 
             double appToRobotYaw(double yaw, std::string unit = "deg");
 
-            double robotToAppYaw(double yaw, std::string unit = "deg");
+            int robotToAppYaw(double yaw, std::string unit = "deg");
 
             int stopRobotMoving();
